# banking-api/src/main/resources/swagger-ui.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /api/v1/accounts:
    get:
      tags:
        - account-controller
      operationId: getAccountById
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Account'
    put:
      tags:
        - account-controller
      operationId: updateAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Account'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
    post:
      tags:
        - account-controller
      operationId: createAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Account'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /api/v1/transactions/{transactionType}:
    post:
      tags:
        - transaction-controller
      operationId: createTransaction
      parameters:
        - name: amount
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: transactionType
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
  /api/v1/transactions/transferfunds:
    post:
      tags:
        - transaction-controller
      operationId: transferFunds
      parameters:
        - name: amount
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: receiverAccountId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: boolean
  /api/v1/transactions:
    get:
      tags:
        - transaction-controller
      operationId: getTransactionsByAccountId
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
  /api/v1/balance:
    get:
      tags:
        - balance-controller
      operationId: findAll
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    $ref: '#/components/schemas/Balance'
  /api/v1/balance/getBalance:
    get:
      tags:
        - balance-controller
      operationId: getBalanceByAccountId
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: number
                format: double
  /api/v1/accounts/role:
    get:
      tags:
        - account-controller
      operationId: getRole
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /api/v1/accounts/all:
    get:
      tags:
        - account-controller
      operationId: getAllAccounts
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
  /api/v1/accounts/{accountId}:
    delete:
      tags:
        - account-controller
      operationId: deleteAccount
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
components:
  schemas:
    Account:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        password:
          type: string
        name:
          type: string
        email:
          type: string
        address:
          type: string
        roles:
          type: string
    Transaction:
      type: object
      properties:
        transaction_id:
          type: integer
          format: int64
        date:
          type: string
          format: date
        amount:
          type: number
          format: double
        transactionType:
          type: string
        description:
          type: string
    Balance:
      type: object
      properties:
        balance_id:
          type: integer
          format: int64
        amount:
          type: number
          format: double
        date:
          type: string
          format: date
        balanceType:
          type: string